# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamDeck API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  name: string;
  icon: string;
  isDefault: boolean;
  createdAt: string;
}

export interface CreateProfileBody {
  name: string;
  icon: string;
  isDefault?: boolean;
}

export interface Folder {
  id: number;
  profileId: number;
  name: string;
  color: string;
  icon: string;
  position: number;
}

export interface CreateFolderBody {
  profileId: number;
  name: string;
  color: string;
  icon: string;
  position: number;
}

export type ButtonActionType =
  (typeof ButtonActionType)[keyof typeof ButtonActionType];

export const ButtonActionType = {
  url: "url",
  hotkey: "hotkey",
  script: "script",
  vpn: "vpn",
  steam: "steam",
  app: "app",
  media: "media",
  obs: "obs",
  github: "github",
  twitch: "twitch",
  zoom: "zoom",
  discord: "discord",
  spotify: "spotify",
  slack: "slack",
  teams: "teams",
  telegram: "telegram",
  notion: "notion",
  browser: "browser",
  system: "system",
  googlemeet: "googlemeet",
  vscode: "vscode",
  youtube: "youtube",
  gmail: "gmail",
  whatsapp: "whatsapp",
  figma: "figma",
  x: "x",
  chatgpt: "chatgpt",
  airdrop: "airdrop",
} as const;

export interface Button {
  id: number;
  profileId: number;
  folderId?: number | null;
  label: string;
  icon: string;
  color: string;
  actionType: ButtonActionType;
  actionValue: string;
  position: number;
  executeCount: number;
  lastExecutedAt?: string | null;
}

export type CreateButtonBodyActionType =
  (typeof CreateButtonBodyActionType)[keyof typeof CreateButtonBodyActionType];

export const CreateButtonBodyActionType = {
  url: "url",
  hotkey: "hotkey",
  script: "script",
  vpn: "vpn",
  steam: "steam",
  app: "app",
  media: "media",
  obs: "obs",
  github: "github",
  twitch: "twitch",
  zoom: "zoom",
  discord: "discord",
  spotify: "spotify",
  slack: "slack",
  teams: "teams",
  telegram: "telegram",
  notion: "notion",
  browser: "browser",
  system: "system",
  googlemeet: "googlemeet",
  vscode: "vscode",
  youtube: "youtube",
  gmail: "gmail",
  whatsapp: "whatsapp",
  figma: "figma",
  x: "x",
  chatgpt: "chatgpt",
  airdrop: "airdrop",
} as const;

export interface CreateButtonBody {
  profileId: number;
  folderId?: number | null;
  label: string;
  icon: string;
  color: string;
  actionType: CreateButtonBodyActionType;
  actionValue: string;
  position: number;
}

export interface ExecutionResult {
  success: boolean;
  message: string;
  buttonId: number;
}

export interface ActivityEntry {
  id: number;
  buttonId: number;
  buttonLabel: string;
  actionType: string;
  success: boolean;
  message: string;
  executedAt: string;
}

export type StatsTopButtonsItem = {
  id: number;
  label: string;
  executeCount: number;
};

export interface Stats {
  totalButtons: number;
  totalProfiles: number;
  totalExecutions: number;
  topButtons: StatsTopButtonsItem[];
}

export type ListButtonsParams = {
  folderId?: number | null;
};