# lib/api-zod/src/generated/types/button.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamDeck API
 * OpenAPI spec version: 0.1.0
 */
import type { ButtonActionType } from "./buttonActionType";

export interface Button {
  id: number;
  profileId: number;
  folderId?: number | null;
  label: string;
  icon: string;
  color: string;
  actionType: ButtonActionType;
  actionValue: string;
  position: number;
  executeCount: number;
  lastExecutedAt?: Date | null;
}