# lib/api-zod/src/generated/types/createFolderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamDeck API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateFolderBody {
  profileId: number;
  name: string;
  color: string;
  icon: string;
  position: number;
}